made it into classes for easy installation and usage
